chore: Update srl

See merge request locker/api-core!522

+++ locker_server/api/v1_0/auto_verity/serializers.py
@@ -9,7 +9,7 @@ class CreateAutoVerifySerializer(serializers.Serializer):
 
 class DeviceAutoVerifySerializer(serializers.Serializer):
     device_id = serializers.CharField(max_length=64)
-    ts = serializers.FloatField()
+    ts = serializers.IntegerField()
     s = serializers.CharField(max_length=128)
     pk = serializers.CharField(max_length=1024)
